WIP: more generic list store
Don't like providing an unwieldy generic list interface,
but if people understand it it can be the right choice.

No tests for generic list methods, but the fallback tests
work, so at least the generic list_split is working
with the new vtable and fallback implementation.

Needed for the generic composite resource list.