Change csalt_resource_use signature
Refactors to use an out parameter store instead of
returning a heap.

Originally returned a void pointer, which would be unmanaged;
then returned a csalt_heap, which would have needed to be
allocated before returning and freed by the caller, which
would have been similarly unmanaged but with extra steps.

An out parameter allows the caller to use an unmanaged store,
like stack memory or global application memory; or a heap
resource, which is now possible to manage with resource_use;
or anything else implementing the store interface.

I normally don't like out parameters but, in this case,
I think it's the best solution. Flexible and promots using
existing solutions to the problem.