Implementation consistent with rest of library
Non-blocking forced on (re-visit this sometime); read/recv
of disconnected socket returns -1 (consistency - -1 means we
won't ever be able to read from this for some reason);
read/recv with EAGAIN/EWOULDBLOCK returns 0 (consistency -
if we expect to be able to read again, but get 0 bytes this
time, we return 0 bytes read).

Need to double-check documentation on this stuff.