Move const declaration to type, remove from funcs
The const declaration in funcs is misleading: it's not
the struct containing the vtable pointer that's const,
it's the pointed-to vtable.

Makes me wonder if these functions _should_ take void*:
then the pointed-to struct containing the vtable pointer
could be made const by const void*, while others can
receive void* to indicate it can be changed.

On the other hand this whole thing came up because I'm
writing a function implementation that _does_ modify the
pointed-to struct, when the original function signature
indicated that the struct would be const.

So like all major decisions in my life, I'm going to
put it off until it's too late.