Introduce CSALT_USE macro
Gives the benefit of providing an easier interface
for running a specific function, at the cost of that
function being less re-usable across the library
(e.g. in csalt_store_split()).

I think the fact that I'm reaching trade-offs on
the simplicity of different implementations suggests
that I'm getting close to "as simple as possible".